refactoring part 2
