+++ Sources/RSocketCore/Extensions/Data.swift
@@ -18,14 +18,14 @@ import Foundation
 import NIO
 import NIOFoundationCompat
 
-protocol DataEncoderProtocol: MultiDataEncoderProtocol {
+public protocol DataEncoderProtocol: MultiDataEncoderProtocol {
     associatedtype Data
     var mimeType: MIMEType { get }
     func encode(_ data: Data, into buffer: inout ByteBuffer) throws
 }
 
 extension DataEncoderProtocol {
-    func encodeIfSupported(_ data: Data, as mimeType: MIMEType, into buffer: inout ByteBuffer) throws -> Bool {
+    public func encodeIfSupported(_ data: Data, as mimeType: MIMEType, into buffer: inout ByteBuffer) throws -> Bool {
         guard mimeType == self.mimeType else { return false }
         try encode(data, into: &buffer)
         return true
@@ -47,8 +47,8 @@ public protocol DataDecoderProtocol: MultiDataDecoderProtocol {
 }
 
 extension DataDecoderProtocol {
-    var supportedMIMETypes: [MIMEType] { [mimeType] }
-    func decodeMIMETypeIfSupported(_ mimeType: MIMEType, from buffer: inout ByteBuffer) throws -> Data? {
+    public var supportedMIMETypes: [MIMEType] { [mimeType] }
+    public func decodeMIMETypeIfSupported(_ mimeType: MIMEType, from buffer: inout ByteBuffer) throws -> Data? {
         guard mimeType == self.mimeType else { return nil }
         return try decode(from: &buffer)
     }
@@ -65,54 +65,50 @@ extension DataDecoderProtocol {
     }
 }
 
-struct JSONDataDecoder<Data: Decodable>: DataDecoderProtocol {
-    var decoder: JSONDecoder = .init()
-    var mimeType: MIMEType { .json }
-    init(type: Data.Type = Data.self, decoder: JSONDecoder = .init()) {
+public struct JSONDataDecoder<Data: Decodable>: DataDecoderProtocol {
+    let decoder: JSONDecoder
+    public var mimeType: MIMEType { .json }
+    public init(type: Data.Type = Data.self, decoder: JSONDecoder = .init()) {
         self.decoder = decoder
     }
-    func decode(from buffer: inout ByteBuffer) throws -> Data {
+    public func decode(from buffer: inout ByteBuffer) throws -> Data {
         try decoder.decode(Data.self, from: buffer)
     }
 }
 
-extension DataDecoderProtocol {
-    static func json(type: Data.Type, decoder: JSONDecoder = .init()) -> JSONDataDecoder<Data> where Data: Decodable {
-            .init(decoder: decoder)
-    }
-}
-
-struct JSONDataEncoder<Data: Encodable>: DataEncoderProtocol {
-    var encoder: JSONEncoder
-    var mimeType: MIMEType { .json }
-    init(type: Data.Type = Data.self, encoder: JSONEncoder = .init()) {
+public struct JSONDataEncoder<Data: Encodable>: DataEncoderProtocol {
+    let encoder: JSONEncoder
+    public var mimeType: MIMEType { .json }
+    public init(type: Data.Type = Data.self, encoder: JSONEncoder = .init()) {
         self.encoder = encoder
     }
-    func encode(_ data: Data, into buffer: inout ByteBuffer) throws {
+    public func encode(_ data: Data, into buffer: inout ByteBuffer) throws {
         try encoder.encode(data, into: &buffer)
     }
 }
 
-enum DataEncoders {
-    struct Map<Encoder: DataEncoderProtocol, Data>: DataEncoderProtocol {
-        var encoder: Encoder
-        var transform: (Data) -> Encoder.Data
-        var mimeType: MIMEType { .json }
-        func encode(_ data: Data, into buffer: inout ByteBuffer) throws {
+public enum DataEncoders {}
+
+extension DataEncoders {
+    public struct Map<Encoder: DataEncoderProtocol, Data>: DataEncoderProtocol {
+        let encoder: Encoder
+        let transform: (Data) -> Encoder.Data
+        public var mimeType: MIMEType { .json }
+        public func encode(_ data: Data, into buffer: inout ByteBuffer) throws {
             try encoder.encode(transform(data), into: &buffer)
         }
     }
-    struct TryMap<Encoder: DataEncoderProtocol, Data>: DataEncoderProtocol {
-        var encoder: Encoder
-        var transform: (Data) throws -> Encoder.Data
-        var mimeType: MIMEType { .json }
-        func encode(_ data: Data, into buffer: inout ByteBuffer) throws {
+    public struct TryMap<Encoder: DataEncoderProtocol, Data>: DataEncoderProtocol {
+        let encoder: Encoder
+        let transform: (Data) throws -> Encoder.Data
+        public var mimeType: MIMEType { .json }
+        public func encode(_ data: Data, into buffer: inout ByteBuffer) throws {
             try encoder.encode(try transform(data), into: &buffer)
         }
     }
 }
 
-extension DataEncoderProtocol {
+public extension DataEncoderProtocol {
     func map<NewData>(
         _ transform: @escaping (NewData) -> Data
     ) -> DataEncoders.Map<Self, NewData> {
@@ -125,26 +121,28 @@ extension DataEncoderProtocol {
     }
 }
 
-enum DataDecoders {
+public enum DataDecoders { }
+    
+public extension DataDecoders {
     struct Map<Decoder: DataDecoderProtocol, Data>: DataDecoderProtocol {
-        var decoder: Decoder
-        var transform: (Decoder.Data) -> Data
-        var mimeType: MIMEType { decoder.mimeType }
-        func decode(from buffer: inout ByteBuffer) throws -> Data {
+        let decoder: Decoder
+        let transform: (Decoder.Data) -> Data
+        public var mimeType: MIMEType { decoder.mimeType }
+        public func decode(from buffer: inout ByteBuffer) throws -> Data {
             transform(try decoder.decode(from: &buffer))
         }
     }
     struct TryMap<Decoder: DataDecoderProtocol, Data>: DataDecoderProtocol {
-        var decoder: Decoder
-        var transform: (Decoder.Data) throws -> Data
-        var mimeType: MIMEType { decoder.mimeType }
-        func decode(from buffer: inout ByteBuffer) throws -> Data {
+        let decoder: Decoder
+        let transform: (Decoder.Data) throws -> Data
+        public var mimeType: MIMEType { decoder.mimeType }
+        public func decode(from buffer: inout ByteBuffer) throws -> Data {
             try transform(try decoder.decode(from: &buffer))
         }
     }
 }
 
-extension DataDecoderProtocol {
+public extension DataDecoderProtocol {
     func map<NewData>(
         _ transform: @escaping (Data) -> NewData
     ) -> DataDecoders.Map<Self, NewData> {
@@ -157,7 +155,7 @@ extension DataDecoderProtocol {
     }
 }
 
-protocol MultiDataEncoderProtocol {
+public protocol MultiDataEncoderProtocol {
     associatedtype Data
     /// tries to encode `data` as `mimeType`.
     /// - Returns: true if it `mimeType` is a supported encoding and encoding was successful
@@ -180,14 +178,14 @@ extension MultiDataEncoderProtocol {
     }
 }
 
-struct MultiDataEncoderTuple2<A, B>: MultiDataEncoderProtocol where
+public struct MultiDataEncoderTuple2<A, B>: MultiDataEncoderProtocol where
 A: DataEncoderProtocol,
 B: DataEncoderProtocol,
 A.Data == B.Data
 {
-    typealias Data = A.Data
+    public typealias Data = A.Data
     let encoder: (A, B)
-    func encodeIfSupported(_ data: Data, as mimeType: MIMEType, into buffer: inout ByteBuffer) throws -> Bool {
+    public func encodeIfSupported(_ data: Data, as mimeType: MIMEType, into buffer: inout ByteBuffer) throws -> Bool {
         switch mimeType {
         case encoder.0.mimeType:
             try encoder.0.encode(data, into: &buffer)
@@ -200,16 +198,16 @@ A.Data == B.Data
     }
 }
 
-struct MultiDataEncoderTuple3<A, B, C>: MultiDataEncoderProtocol where
+public struct MultiDataEncoderTuple3<A, B, C>: MultiDataEncoderProtocol where
 A: DataEncoderProtocol,
 B: DataEncoderProtocol,
 C: DataEncoderProtocol,
 A.Data == B.Data,
 A.Data == C.Data
 {
-    typealias Data = A.Data
+    public typealias Data = A.Data
     let encoder: (A, B, C)
-    func encodeIfSupported(_ data: Data, as mimeType: MIMEType, into buffer: inout ByteBuffer) throws -> Bool {
+    public func encodeIfSupported(_ data: Data, as mimeType: MIMEType, into buffer: inout ByteBuffer) throws -> Bool {
         switch mimeType {
         case encoder.0.mimeType:
             try encoder.0.encode(data, into: &buffer)
@@ -225,7 +223,7 @@ A.Data == C.Data
 }
 
 @resultBuilder
-enum MultiDataEncoderBuilder {
+public enum MultiDataEncoderBuilder {
     static func buildBlock<Encoder>(
         _ encoder: Encoder
     ) -> Encoder where Encoder: MultiDataEncoderProtocol {
@@ -252,14 +250,10 @@ public protocol MultiDataDecoderProtocol {
     func decodeMIMETypeIfSupported(_ mimeType: MIMEType, from buffer: inout ByteBuffer) throws -> Data?
 }
 
-enum MultiDataDecoderError: Swift.Error {
-    case mimeTypeNotSupported(MIMEType)
-}
-
 extension MultiDataDecoderProtocol {
     func decodeMIMEType(_ mimeType: MIMEType, from buffer: inout ByteBuffer) throws -> Data {
         guard let data = try decodeMIMETypeIfSupported(mimeType, from: &buffer) else {
-            throw MultiDataDecoderError.mimeTypeNotSupported(mimeType)
+            throw Error.invalid(message: "\(mimeType) not supported")
         }
         return data
     }
@@ -276,16 +270,16 @@ extension MultiDataDecoderProtocol {
     }
 }
 
-struct MultiDataDecoderTuple2<A, B>: MultiDataDecoderProtocol where
+public struct MultiDataDecoderTuple2<A, B>: MultiDataDecoderProtocol where
 A: DataDecoderProtocol,
 B: DataDecoderProtocol,
 A.Data == B.Data {
-    typealias Data = A.Data
+    public typealias Data = A.Data
     let decoder: (A, B)
-    var supportedMIMETypes: [MIMEType] {
+    public var supportedMIMETypes: [MIMEType] {
         [decoder.0.mimeType, decoder.1.mimeType]
     }
-    func decodeMIMETypeIfSupported(
+    public func decodeMIMETypeIfSupported(
         _ mimeType: MIMEType,
         from buffer: inout ByteBuffer
     ) throws -> Data? {
@@ -300,18 +294,18 @@ A.Data == B.Data {
     }
 }
 
-struct MultiDataDecoderTuple3<A, B, C>: MultiDataDecoderProtocol where
+public struct MultiDataDecoderTuple3<A, B, C>: MultiDataDecoderProtocol where
 A: DataDecoderProtocol,
 B: DataDecoderProtocol,
 C: DataDecoderProtocol,
 A.Data == B.Data,
 A.Data == C.Data{
-    typealias Data = A.Data
+    public typealias Data = A.Data
     let decoder: (A, B, C)
-    var supportedMIMETypes: [MIMEType] {
+    public var supportedMIMETypes: [MIMEType] {
         [decoder.0.mimeType, decoder.1.mimeType, decoder.2.mimeType]
     }
-    func decodeMIMETypeIfSupported(
+    public func decodeMIMETypeIfSupported(
         _ mimeType: MIMEType,
         from buffer: inout ByteBuffer
     ) throws -> Data? {
@@ -329,7 +323,7 @@ A.Data == C.Data{
 }
 
 @resultBuilder
-enum MultiDataDecoderBuilder {
+public enum MultiDataDecoderBuilder {
     static func buildBlock<Decoder>(
         _ decoder: Decoder
     ) -> Decoder where Decoder: MultiDataDecoderProtocol {
